refactor: move app's files in app folder
