+++ rs-react-app/src/app/App.test.ts
@@ -2,14 +2,14 @@ import { screen, waitFor, fireEvent } from '@testing-library/react';
 import { describe, it, expect, vi } from 'vitest';
 import '@testing-library/jest-dom';
 import '@testing-library/dom';
+import { mockSuccessConfig } from '../../test-utils/mocks/resultBlockMock';
+import * as fetchModule from './../service/request';
+import RouterComponent from './../service/router/RouterComponent';
+import { renderWithRouter } from '../../test-utils/renderWithRouter';
 import {
-  windowClear,
   localStorageMock,
-} from '../test-utils/mocks/localStorage';
-import { mockSuccessConfig } from '../test-utils/mocks/resultBlockMock';
-import * as fetchModule from './service/request';
-import RouterComponent from './service/router/RouterComponent';
-import { renderWithRouter } from '../test-utils/renderWithRouter';
+  windowClear,
+} from '../../test-utils/mocks/localStorage';
 
 windowClear();
